refactor(language-server): modularize diagnostics orchestration
