+++ language-server/src/diagnostics/constants.ts
@@ -1,0 +1,1 @@
+export const DTIF_DIAGNOSTIC_SOURCE = 'dtif-schema';
