+++ language-server/src/diagnostics/index.ts
@@ -1,0 +1,2 @@
+export { DocumentValidator, type DocumentValidatorOptions } from './document-validator.js';
+export type { DtifDiagnosticData } from './types.js';
