+++ language-server/src/index.ts
@@ -6,4 +6,4 @@ export {
   type ManagedDocuments
 } from './server.js';
 export { buildInitializeResult, type DtifInitializeResult } from './runtime/initialize.js';
-export { DocumentValidator } from './diagnostics.js';
+export { DocumentValidator } from './diagnostics/index.js';
